(Feat): Moved job processing logic from api.py to worker.py [tsv22][at2222]
