+++ aignostic/Dockerfile.worker
@@ -1,0 +1,6 @@
+FROM python:3.12
+
+# INSTALL PIKA
+
+RUN pip install pika
+
