+++ aignostic/pyproject.toml
@@ -34,6 +34,7 @@ pandas-stubs = "^2.2.3.241126"
 pytest = "^8.3.4"
 pytest-asyncio = "^0.25.2"
 update = "^0.0.1"
+pika = "^1.3.2"
 
 [tool.poetry.group.dev.dependencies]
 autopep8 = "2.3.1"
